OMR AOT: Introducing AOTMethodHeader class hierarchy
This commit brings in the following files: OMRAotMethodHeader,
OMRAotMethodHeader.cpp, AOTMethodHeader.cpp. The AOTMethodHeader files
are extensible class hierarchy proposed to be used for serialization and
deserialization of relocations buffer to and from the storage, alongside
of the compiled code. This class is supposed to be extended for caching
more data on the level of compiled code.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>