OMR AOT: Addition of RelocationRuntime, RelocationTarget
The RelocationRuntime and a class extending it, OMRAOTRelocation Runtime
as well as RelocationTarget classes are ported from OpenJ9 into OMR.
RelocationRuntime is responsible for temporarily storing relocations at
the stage of their creation (at compile time). During the load-time the
relocation runtime creates relocationRecord groups that iterate over
relocation buffer and apply the actual relocations. RelocationRuntime
also holds reference to an architecture specific relocationTarget
(that was transformed into an extensible class). The relocation target
is a class that implements an API to store objects of different size.
Having this class allows for abstracting away from memory implementation
of an particular device and might potentially improve portability of OMR
for embedded devices.The TRRelocationRuntime.hpp and
TRRelocationTarget.hpp are the header files required for implementing
concrete versions of extensible class hierarchy in OMR, by defining
their version in a TR namespace. The OMRAOTRelocationRuntime holds a
lookup by name map that would allow relocation runtime to have named
symbol addresses and would allow linking external functions into runtime.
The keys being strings is an intermediate in-development solution is
supposed to change into something more suitable for a method
identification, Resolved Method being a candidate for replacement. For
now, maintaining the string being sufficient for a method verification
is left as the runtime constructor responsibility.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>