OMR AOT: AOTAdapter class hierarchy
The AOTAdapter class is supposed to provide
This commit introduces the OMRAotAdapter, an extensible class that is
designed to dispatch the interaction between compiler components to
perform persisting of relocatable code. The benefit of having this
class is in decoupling aot-related classes and removing indirect
responsibility by introducing new layer of abstraction: the
RelocationRuntime class is only concerned about managing generation
and application of relocations, without knowing about persisting
method implemented in current runtime environment. Similarly, Shared
Cache or other options for persisting code should implement simplistic
interface (that, obviously can be extended) that would have two
functions: receive a message that contains a serialized version of
AOTMethod header, as well as send one. The AOTAdapter will register
the header and thus make it available to RelocationRuntime and hence,
the compiler, as well as to runtime environment, since AOTAdapter is
planned to be alive longer than a compilation thread. Currently, the
methods are identified by const char* , later to be replaced with proper
calls to ResolvedMethod API, that would allow to generate method signa-
tures for verification

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>