Unify doRegisterAssignment on all platforms except x86
Now that we have gotten `doRegisterAssignment` to look exactly the same
on several platforms we can deprecte the platform specific versions.
Unfortunately on x86 we seem to be doing something completely different
from all the other platforms, so in general tackling register allocation
on x86 will need to be done separately.