Allow 64K page size as a valid option on AIX
xlpCodeCacheTests fails on AIX because the default
code cache large page size 64K is absent in the
list of available code cache large page sizes
output by `-verbose:sizes`.

Include 64K in the list of available code cache
large page sizes for the test to parse as the
expected page size when no `-Xlp` option is
supplied and the default page size is used.

Signed-off-by: Amarpreet Singh amarpreet1997@gmail.com