OMR AOT: AOTLoadStoreDriver class hierarchy
This commit introduces the OMRAOTLoadStoreDriver, an extensible class
that is designed to dispatch the interaction between compiler components
to perform persisting of relocatable code. The benefit of having this
class is in decoupling aot-related classes and removing indirect
responsibility by introducing new layer of abstraction: the
RelocationRuntime class is only concerned about managing generation
and application of relocations, without knowing about persisting
method implemented in current runtime environment. Similarly, Shared
Cache or other options for persisting code should implement simplistic
interface (that, obviously can be extended) that would have two
functions: receive a message that contains a serialized version of
AOTMethod header, as well as send one. The AOTLoadStoreDriver will
register the header and thus make it available to RelocationRuntime and
hence, the compiler, as well as to runtime environment, since
AOTLoadStoreDriver is planned to be alive longer than a compilation
thread. Currently, the methods are identified by const char* ,
later to be replaced with proper calls to ResolvedMethod API, that would
allow to generate method signatures for verification.
AOTLoadStoreDriver alos holds a lookup by name map that would allow
relocations to have named symbol addresses and would allow linking
external functions into runtime. The keys being strings is an intermedi-
ate in-development solution is supposed to change into something more
suitable for a method identification, Resolved Method being a candidate
for replacement. For now, maintaining the string being sufficient for a
method verification is left as the runtime constructor responsibility.

Signed-off-by: Georgiy Krylov <georgiy.krylov@unb.ca>